Cast to __int128 just in case.

+++ test/mulmod_test.cpp
@@ -239,7 +239,7 @@ TEST( MulmodTest, mulmod_barrett ) {
   for( uint64_t i = 0; i < NUM_TESTS; ++i ) {
     int prime_modulus = test_data[i][3];
     int n = n_clog( prime_modulus );
-    __int128 numerator = 1 << (n * 2);
+    __int128 numerator = (__int128) 1 << (n * 2);
     uint64_t mu = numerator / prime_modulus;
     long double r = (long double) mu / (1 << n);
     ASSERT_EQ( test_data[i][0], mulmod_barrett( test_data[i][1], test_data[i][2], prime_modulus, n, r ) );
